refactor!: CommentProcessorRegistry class location and update module exports
The CommentProcessorRegistry class has been moved from the org.wickedsource.docxstamper.processor package to the pro.verron.officestamper.core package. The relevant exports in module-info have been updated to reflect this reorganization of code for better package structure and separation of concerns. Moreover, unnecessary import statements have been removed, thus, keep the code clean and concise.
